Honorarium Amounts Updated

+++ IndiaEvents.Models/Models/EventTypeSheets/HonorariumPaymentP2.cs
@@ -19,10 +19,10 @@ namespace IndiaEvents.Models.Models.EventTypeSheets
         public string? StartTime { get; set; }
         public string? EndTime { get; set; }
         public string? VenueName { get; set; }
-        public int? TotalTravelAndAccomodationSpend { get; set; }
-        public int? TotalHonorariumSpend { get; set; }
-        public int? TotalSpend { get; set; }
-        public int? TotalLocalConveyance { get; set; }
+        public double? TotalTravelAndAccomodationSpend { get; set; }
+        public double? TotalHonorariumSpend { get; set; }
+        public double? TotalSpend { get; set; }
+        public double? TotalLocalConveyance { get; set; }
         public string? Brands { get; set; }
         public string? Invitees { get; set; }
         public string? Panelists { get; set; }
@@ -42,9 +42,9 @@ namespace IndiaEvents.Models.Models.EventTypeSheets
         public string? FinanceTreasury { get; set; }
         public string? SlideKits { get; set; }
         public string? Expenses { get; set; }
-        public int? TotalTravelSpend { get; set; }
-        public int? TotalAccomodationSpend { get; set; }
-        public int? TotalExpenses { get; set; }
+        public double? TotalTravelSpend { get; set; }
+        public double? TotalAccomodationSpend { get; set; }
+        public double? TotalExpenses { get; set; }
         public string? IsDeviationUpload { get; set; }
 
 
@@ -60,7 +60,7 @@ namespace IndiaEvents.Models.Models.EventTypeSheets
         public string? MisCode { get; set; }
         public string? GOorNGO { get; set; }
         public string? IsInclidingGst { get; set; }
-        public int? AgreementAmount { get; set; }
+        public double? AgreementAmount { get; set; }
         public string? IsAnnualTrainerAgreementValid { get; set; }
 
     }
